// src/design-system/tokens/generated/tokens.ts
export const tokens = {
  "colors": {
// ...
    "grey": {
      "100": {
        "value": "#E5E5E5",
        "filePath": "src/design-system/tokens/source/colors.json",
        "isSource": true,
        "original": {
          "value": "#E5E5E5"
        },
        "name": "ColorGrey100",
        "attributes": {
          "category": "color",
          "type": "grey",
          "item": "100"
        },
        "path": [
          "color",
          "grey",
          "100"
        ]
      },
      "200": {
        "value": "#A3A3A3",
        "filePath": "src/design-system/tokens/source/colors.json",
        "isSource": true,
        "original": {
          "value": "#A3A3A3"
        },
        "name": "ColorGrey200",
        "attributes": {
          "category": "color",
          "type": "grey",
          "item": "200"
        },
        "path": [
          "color",
          "grey",
          "200"
        ]
      },
      "300": {
        "value": "#737373",
        "filePath": "src/design-system/tokens/source/colors.json",
        "isSource": true,
        "original": {
          "value": "#737373"
        },
        "name": "ColorGrey300",
        "attributes": {
          "category": "color",
          "type": "grey",
          "item": "300"
        },
        "path": [
          "color",
          "grey",
          "300"
        ]
      },
      "400": {
        "value": "#525252",
        "filePath": "src/design-system/tokens/source/colors.json",
        "isSource": true,
        "original": {
          "value": "#525252"
        },
        "name": "ColorGrey400",
        "attributes": {
          "category": "color",
          "type": "grey",
          "item": "400"
        },
        "path": [
          "color",
          "grey",
          "400"
        ]
      },
      "500": {
        "value": "#404040",
        "filePath": "src/design-system/tokens/source/colors.json",
        "isSource": true,
        "original": {
          "value": "#404040"
        },
        "name": "ColorGrey500",
        "attributes": {
          "category": "color",
          "type": "grey",
          "item": "500"
        },
        "path": [
          "color",
          "grey",
          "500"
        ]
      },
      "600": {
        "value": "#262626",
        "filePath": "src/design-system/tokens/source/colors.json",
        "isSource": true,
        "original": {
          "value": "#262626"
        },
        "name": "ColorGrey600",
        "attributes": {
          "category": "color",
          "type": "grey",
          "item": "600"
        },
        "path": [
          "color",
          "grey",
          "600"
        ]
      }
    },
// ...
  },
// ...
  "border": {
    "solid": {
      "1": {
        "value": 1,
        "type": "border",
        "filePath": "src/design-system/tokens/source/borders.json",
        "isSource": true,
        "original": {
          "value": 1,
          "type": "border"
        },
        "name": "BorderSolid1",
        "attributes": {
          "category": "border",
          "type": "solid",
          "item": "1"
        },
        "path": [
          "border",
          "solid",
          "1"
        ]
      },
      "2": {
        "value": 2,
        "type": "border",
        "filePath": "src/design-system/tokens/source/borders.json",
        "isSource": true,
        "original": {
          "value": 2,
          "type": "border"
        },
        "name": "BorderSolid2",
        "attributes": {
          "category": "border",
          "type": "solid",
          "item": "2"
        },
        "path": [
          "border",
          "solid",
          "2"
        ]
      },
      "3": {
        "value": 3,
        "type": "border",
        "filePath": "src/design-system/tokens/source/borders.json",
        "isSource": true,
        "original": {
          "value": 3,
          "type": "border"
        },
        "name": "BorderSolid3",
        "attributes": {
          "category": "border",
          "type": "solid",
          "item": "3"
        },
        "path": [
          "border",
          "solid",
          "3"
        ]
      },
      "4": {
        "value": 4,
        "type": "border",
        "filePath": "src/design-system/tokens/source/borders.json",
        "isSource": true,
        "original": {
          "value": 4,
          "type": "border"
        },
        "name": "BorderSolid4",
        "attributes": {
          "category": "border",
          "type": "solid",
          "item": "4"
        },
        "path": [
          "border",
          "solid",
          "4"
        ]
      },
      "5": {
        "value": 5,
        "type": "border",
        "filePath": "src/design-system/tokens/source/borders.json",
        "isSource": true,
        "original": {
          "value": 5,
          "type": "border"
        },
        "name": "BorderSolid5",
        "attributes": {
          "category": "border",
          "type": "solid",
          "item": "5"
        },
        "path": [
          "border",
          "solid",
          "5"
        ]
      }
    },
// ...
  },
// ...
} as const;